Reorganizes the test folder into subdirectories based on resource
