--- tests/api_tests/test_session.py
@@ -1,20 +1,0 @@
-import json
-import pytest
-
-#imports test data
-from tests.data.contact_data import CONTACTS_API, EMAILS_API
-from tests.data.skill_data import CONTACT_SKILLS
-from tests.data.profile_data import PROFILES_API
-from tests.data.program_data import PROGRAM_APPS_API
-
-
-class TestSession:
-
-    def test_get(self):
-        assert 1
-
-    def test_post(self):
-        assert 1
-
-    def test_delete(self):
-        assert 1
