--- tests/data/testing_data.py
@@ -1,6 +1,0 @@
-
-
-INSTRUCTIONS = {
-    'billy': {},
-    'obama': {},
-}
