+++ tests/program/program_data.py
@@ -1,4 +1,4 @@
-from tests.data.contact_data import CONTACTS_API
+from tests.contact.contact_data import CONTACTS_API
 
 PROGRAMS_DATABASE = {
     'pfp': {
