+++ tests/program/test_program.py
@@ -1,0 +1,10 @@
+import json
+import pytest
+
+#imports test data
+
+
+class TestProgramAll:
+
+    def test_get(self):
+        assert 1
