+++ tests/session/test_session.py
@@ -1,0 +1,17 @@
+import json
+import pytest
+
+#imports test data
+
+
+
+class TestSession:
+
+    def test_get(self):
+        assert 1
+
+    def test_post(self):
+        assert 1
+
+    def test_delete(self):
+        assert 1
